deprecate old esri_ascii functionsi, move into deprecated module
