+++ landlab/io/__init__.py
@@ -1,16 +1,16 @@
 """Modules that read/write ModelGrids from various file formats."""
 
-from landlab.io.esri_ascii import BadHeaderLineError
-from landlab.io.esri_ascii import DataSizeError
-from landlab.io.esri_ascii import KeyTypeError
-from landlab.io.esri_ascii import KeyValueError
-from landlab.io.esri_ascii import MismatchGridDataSizeError
-from landlab.io.esri_ascii import MismatchGridXYLowerLeft
-from landlab.io.esri_ascii import MismatchGridXYSpacing
-from landlab.io.esri_ascii import MissingRequiredKeyError
-from landlab.io.esri_ascii import read_asc_header
-from landlab.io.esri_ascii import read_esri_ascii
-from landlab.io.esri_ascii import write_esri_ascii
+from landlab.io._deprecated_esri_ascii import BadHeaderLineError
+from landlab.io._deprecated_esri_ascii import DataSizeError
+from landlab.io._deprecated_esri_ascii import KeyTypeError
+from landlab.io._deprecated_esri_ascii import KeyValueError
+from landlab.io._deprecated_esri_ascii import MismatchGridDataSizeError
+from landlab.io._deprecated_esri_ascii import MismatchGridXYLowerLeft
+from landlab.io._deprecated_esri_ascii import MismatchGridXYSpacing
+from landlab.io._deprecated_esri_ascii import MissingRequiredKeyError
+from landlab.io._deprecated_esri_ascii import read_asc_header
+from landlab.io._deprecated_esri_ascii import read_esri_ascii
+from landlab.io._deprecated_esri_ascii import write_esri_ascii
 from landlab.io.legacy_vtk import write_legacy_vtk
 from landlab.io.obj import write_obj
 from landlab.io.shapefile import read_shapefile
